+++ CHANGES.md
@@ -6,7 +6,11 @@
 
 * Placeholder
 
+
 * `map_record` now returns `nil` if record was skipped.
+
+* Traject::Indexer::Settings no longer has it's own global defaults, Instead it can be given a set of defaults with #with_defaults, usually right after instantiation. To support different defaults for different Indexers.
+
 
 ### Added